==> set_up/3_conservation_features.Rmd <==
---
title: "Conservation Features Set Up"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# Load in packages
library(raster)
library(tidyverse)
library(sf)
library(here)
library(fasterize)
library(janitor)
```

# Critical Habitats

In this Rmarkdown, we are creating our critical habitat final rasters which include: coral, seagrass, mangroves, seamounts, and knolls. This will be added to the `Prioritzr` conservation features along with the species distributions created in `2_iucn_species.Rmd`. Other habitat shapefiles can be used following the same setup.

### **STEP 1: Read in Mozambique EEZ and Shapefile**
Grabbing the project area from the `1_MZ_EEZ_raster.Rmd` with the cell ids and the 100m resolutions. 
```{r}
## read in MZ EEZ tif
mz_rast_id <- raster(here("set_up/final_rasters/mz_eez_templates/mz_rast_id.tif"))

## Grab Moz Exclusive Economic Zone (EEZ) shapefile from local computer
mz_eez_path <- 'G:/group_project/data/mz_eez'
mz_shape <- list.files(mz_eez_path, full.names = TRUE) 

mz_eez_sf <- sf::read_sf(mz_shape[6])

## check coordinate system
st_crs(mz_eez_sf)

## Add a buffer of 10 km to make sure cells on the outer edges of the EEZ are included
mz_eez_buffer_sf <- sf::st_buffer(mz_eez_sf, dist = 10000)

## Grab Moz Exclusive Economic Zone (EEZ)  at 100m shapefile from local computer
mz_eez_100_path <- 'G:/group_project/data/mz_raster'
mz_100_shape <- list.files(mz_eez_100_path, full.names = TRUE) 

mz_rast_100 <- raster(mz_100_shape[1])

```

### **STEP 2: Create all habitat rasters**
The following code goes through how we rasterized each habitat. They all follow similar steps, but in same cases we needed to aggregate the data to be able to view the data. Each finalized habitat raster was saved in the conservation features habitat folder. All raw data was saved on local computer initially so follow personal file path to each data set. More or less habitat can be added if needed. 

#### Read in ACA data for Coral and Seagrass Shapefiles
```{r}
## Grab aca shapefile from local computer
## Path to get data on G shared drive at Bren
aca_path <- 'G:/group_project/Data/Habitats/ACA_classes'
aca_shape <- list.files(aca_path, full.names = TRUE) 

## Read in shapefile as a simple feature and tranform to the correct crs
aca_sf <- sf::read_sf(aca_shape[8])%>% 
  st_transform(., crs = st_crs(mz_eez_sf))

## Check the CRS of the coral sf to make sure it matches the CRS of mz_rast
crs(mz_rast_id)
st_crs(aca_sf)

## Create data frame to easily check data inside sf
aca_df <- aca_sf %>% 
  as.data.frame() %>% 
  select(-geometry) 
```


#### CORAL
```{r}
### coral shapefile
aca_coral_sf <- aca_sf %>% 
  filter(Ecosystem == "Coral Reefs") %>% 
  select(-class)

## Make a raster of the coral simple feature using mz_rast_100 - the raster we created at a smaller resolution to set extent and mask to the mz_eez_buffer_sf
aca_coral_raster <- fasterize::fasterize(aca_coral_sf, mz_rast_100)%>%
  mask(mz_eez_sf)

## plot to make sure it looks good
plot(aca_coral_raster) # look closely it's there!

## Save as raster
writeRaster(aca_coral_raster, here('set_up/final_rasters/conservation_features/coral/aca_coral.tif'), overwrite = TRUE)

# Aggregate to become a coarser resolution at 10000 m
aca_coral_raster_agg <- raster::aggregate(aca_coral_raster, fact=100, fun=sum, na.rm = TRUE)
aca_coral_aggregate_per <- aca_coral_raster_agg/10000

## mask to the MZ EEZ
aca_coral_aggregate_per <- mask(aca_coral_aggregate_per, mz_eez_sf)

## Lets plot to see what it looks like
plot(aca_coral_aggregate_per)

## Save coral aggregate as raster
writeRaster(aca_coral_aggregate_per, here('set_up/final_rasters/conservation_features/coral/aca_coral_agg.tif'), overwrite = TRUE)
```


#### SEAGRASS
```{r}
### seagrass shapefile
aca_seagrass_sf <- aca_sf %>% 
  filter(Ecosystem == "Seagrass") %>% 
  select(-class)

## Let's create a rsater using the mz_rast with resolution of 10000 and mask to the MZ EEZ
aca_seagrass_raster <- fasterize::fasterize(aca_seagrass_sf, mz_rast_100) %>% 
  mask(mz_eez_sf)

## plot to make sure it looks good
plot(aca_seagrass_raster)

## Save as raster
writeRaster(aca_seagrass_raster, here('set_up/final_rasters/conservation_features/seagrass/aca_seagrass.tif'), overwrite = TRUE)

# Aggregate to become a coarser resolution at 10000 m
aca_seagrass_raster_agg <- raster::aggregate(aca_seagrass_raster, fact=100, fun=sum, na.rm = TRUE)
aca_seagrass_aggregate_per <- aca_seagrass_raster_agg/10000

aca_seagrass_aggregate_per <- mask(aca_seagrass_aggregate_per, mz_eez_sf)

plot(aca_seagrass_aggregate_per)

## Save coral aggregate as raster
writeRaster(aca_seagrass_aggregate_per, here('set_up/final_rasters/conservation_features/seagrass/aca_seagrass_agg.tif'), overwrite = TRUE)
```


#### MANGROVES
```{r}
## Path to get data on G shared drive at Bren
mangrove_path <- 'G:/group_project/Data/Habitats/mangroves'
mangrove_shape <- list.files(mangrove_path, full.names = TRUE)

## Read in shapefile as a simple feature and select relevant columns
mangrove_sf <- sf::read_sf(mangrove_shape[6]) %>%
  select("OBJECTID", "ISO3", "AREA_KM2", "AREA_M2", "geometry")

## Plot to see what simple feature looks like
plot(mangrove_sf)

## Check the CRS of the mangrove sf and compare to mz_rast
st_crs(mz_rast_id)
st_crs(mangrove_sf)

## Create data frame to easily check data inside coral sf
mangrove_df <- mangrove_sf %>% as.data.frame() %>% select(-geometry)

## Mangroves also has finer resolution that we want to be sure we capture so lets make a raster of the mangove simple feature using mz_rast_100 - the raster we created at a smaller resolution to set extent and mask to the mz_eez
mangrove_raster <- fasterize::fasterize(mangrove_sf, mz_rast_100)%>%
  mask(mz_eez_sf)

## plot to make sure it looks good
plot(mangrove_raster)

## Now we will aggregate to become a coaser resoltuion that matches our other rasters at 10000 m
mangrove_agg <- raster::aggregate(mangrove_raster, fact=100, fun=sum, na.rm=TRUE)
mangrove_agg_percent <- mangrove_agg/10000

mangrove_rast <- mask(mangrove_agg_percent, mz_eez_sf)

## Lets plot to see what it looks like
plot(mangrove_rast)

## Lookgs good - let's save as a tif file
writeRaster(mangrove_agg_percent, here('set_up/final_rasters/conservation_features/mangrove/mangrove.tif'), overwrite = TRUE)

```

#### SEAMOUNTS
```{r}
## Grab seamount shapefile from the local computer
seamount_path <- 'G:/group_project/Data/Habitats/seamounts'
seamount_shape <- list.files(seamount_path, full.names = TRUE)

## Load shapefile as simple feature
seamount_sf <- sf::read_sf(seamount_shape[6]) 

## Check the CRS of the seamount and compare to the mz rast
st_crs(mz_rast_id)
st_crs(seamount_sf)

## Let's create a rsater using the mz_rast with resolution of 10000 and mask tot the MZ EEZ
seamount_raster <- fasterize::fasterize(seamount_sf, mz_rast_id)%>%
  mask(mz_eez_sf)

## Plot to see what it looks like
plot(seamount_raster)

# Looks good lets save as a tif
writeRaster(seamount_raster, here('set_up/final_rasters/conservation_features/seamounts/seamount.tif'), overwrite = TRUE)
```

#### KNOLLS
```{r}
## Grab knolls shapefile from the local computer
knolls_path <- 'G:/group_project/Data/Habitats/knolls'
knolls_shape <- list.files(knolls_path, full.names = TRUE)

## Load shapefile as simple feature
knolls_sf <- sf::read_sf(knolls_shape[6]) 

## Check the CRS of the knolls sf
st_crs(knolls_sf)

## Let's create a rsater using the mz_rast with resolution of 10000 and mask tot the MZ EEZ
knolls_raster <- fasterize::fasterize(knolls_sf, mz_rast_id)%>%
  mask(mz_eez_sf)

## Plot to see what it looks like
plot(knolls_raster)

# Looks good - lets save as a tif
writeRaster(knolls_raster, here('set_up/final_rasters/conservation_features/knolls/knolls.tif'), overwrite = TRUE)

```

### **STEP 3: Create raster stack**
We need one raster that contains all habitat raster for `Prioritzr` so we create a raster stack of the few we created and save.

```{r}
## Create a features stack from all five critical habitats created in the previous steps
habitat_stack <- stack(aca_coral_raster_agg, mangrove_rast, aca_seagrass_raster_agg, seamount_raster, knolls_raster)

## Plot to see what it looks like
plot(habitat_stack)

plot(habitat_stack, main = paste("Feature", seq_len(nlayers(habitat_stack))),
     nr = 2, box = FALSE, axes = FALSE)

## Looks good, let's save as a multilayer feature tif
writeRaster(habitat_stack, filename=here("set_up","final_rasters","conservation_features", "habitat_stack.tif"), options="INTERLEAVE=BAND", overwrite=TRUE)
```

### **STEP 4: Combine habitat raster stack and species raster stack to create stack of all conservation features**
The conservation features for our model is both the critical habitats and species distributions, so we need to create a raster stack of the two raster stack: `species_stack.tif` & `habitat_stack.tif`. 

```{r}
## pull in species stack we created in step 2
species_stack <- brick(here("set_up", "final_rasters", "conservation_features", "species_stack.tif"))

plot(species_stack)

### stack all species and habitat
features_stack <- stack(habitat_stack, species_stack)

plot(features_stack)

### change into a dataframe to check it
features_df <- as.data.frame(features_stack)

### write raster
writeRaster(features_stack, here('set_up/final_rasters/conservation_features/all_features_stack.tif'),options="INTERLEAVE=BAND", overwrite = TRUE)

```

Outputs:
 -`aca_coral` = raster of coral within Mozambique EEZ at 100m resolution
 -`aca_coral_agg` = raster of coral within Mozambique EEZ aggregated to match 10000m resolution
 -`aca_seagrass` = raster of seagrass within Mozambique EEZ at 100m resolution
 -`aca_seagrass_agg` = raster of seagrass within Mozambique EEZ aggregated to match 10000m resolution
 -`mangrove` = raster of mangrove within Mozambique EEZ at 100m resolution
 -`mangrove_agg` = raster of mangrove within Mozambique EEZ aggregated to match 10000m resolution
 -`knolls` = raster of knolls within Mozambique EEZ at 10000m resolution
 -`seamounts` = raster of seamounts within Mozambique EEZ at 10000m resolution
 -`habitat_stack` = raster stack of all critical habitats


Citations:

Coral Reefs: Jiwei Li, David E. Knapp, Nicholas S. Fabina, Emma V. Kennedy, Kirk Larsen, Mitchell B. Lyons, Nicholas J. Murray, Stuart R. Phinn, Chris M. Roelfsema, and Gregory P. Asner. A global coral reef probability map generated using convolutional neural networks. Coral Reefs. 2020. Advance online publication. https://doi.org/10.1007/s00338-020-02005-6

Mangroves: Giri C, Ochieng E, Tieszen LL, Zhu Z, Singh A, Loveland T, Masek J, Duke N (2011). Status and distribution of mangrove forests of the world using earth observation satellite data (version 1.3, updated by UNEP-WCMC). Global Ecology and Biogeography 20: 154-159. Paper DOI: 10.1111/j.1466-8238.2010.00584.x; Data URL: http://data.unep-wcmc.org/datasets/4

Seamounts and Knolls: Yesson C, Clark MR, Taylor M, Rogers AD (2011). The global distribution of seamounts based on 30-second bathymetry data. Deep Sea Research Part I: Oceanographic Research Papers 58: 442-453. DOI: http://dx.doi.org/10.1016/j.dsr.2011.02.004. Data DOI: http://doi.pangaea.de/10.1594/PANGAEA.757564.
 

End script.
