==> set_up/6_industrial_layer_setup.Rmd <==
---
title: "Industrial input file"
author: "Vanessa Rathbone"
date: "1/19/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#load packages
library(raster)
library(tidyverse)
library(sf)
library(here)
library(fasterize)
library(stringr)
library(janitor)
library(fasterize)
library(tmap)
```

# Cost: Industrial Fishing Pressure

This is the first of many different cost layers we added to our prioritization model. Industrial fishing is a major threat to shark and ray species largely due to bycatch, so we wanted to include it in our analysis. This layer was created using VMS data provided by the Mozambique government as a proxy for industrial fishing effort. The VMS data was point data from vessels across four industrial fisheries - long-line, shrimp, gamba, and tuna. The point data were smoothed into a kernel density raster for each fishing gear. Since the VMS data were not comprehensive, the kernel density surface was used to interpolate values where there are no fishing points within a search radius. In this way, areas with lots of fishing points receive high values, areas with fewer points receive low values, and areas beyond a 20-km search radius receive values of 0. We log transformed each of the industrial fishing layers and rescaled them between 0 and 1 to put them on a unitless scale that allows direct comparison across cost layers. We then combined them into one cumulative industrial fishing pressure layer and rescaled the cumulative layer between 0 and 1.

#### **STEP 1: Read in Mozambique EEZ and Shapefile**
Grabbing the project area from the `1_MZ_EEZ_raster.Rmd` with the cell ids and the shapefiles.

```{r}
## read in MZ EEZ tif
mz_rast_id <- raster(here("set_up/final_rasters/mz_eez_templates/mz_rast_id.tif"))

## Grab Moz Exclusive Economic Zone (EEZ) shapefile from local computer
mz_eez_path <- 'G:/group_project/data/mz_eez'
mz_shape <- list.files(mz_eez_path, full.names = TRUE) 

mz_eez_sf <- sf::read_sf(mz_shape[6])

## check coordinate system
st_crs(mz_eez_sf)

## Add a buffer of 10 km to make sure cells on the outer edges of the EEZ are included
mz_eez_buffer_sf <- sf::st_buffer(mz_eez_sf, dist = 10000)
```

#### **STEP 2: Read in all industrial rasters**
```{r}
#read in tifs as rasters
long_line <- raster(here("set_up", "raw_data", "industrial", "fishing_line_kernel_density_noports.tif"))

gamba <- raster(here("set_up", "raw_data", "industrial", "Gamba_fishing_density_noports.tif"))

shrimp <- raster(here("set_up", "raw_data", "industrial", "Shrimp_fishing_density_noports.tif"))

tuna <- raster(here("set_up", "raw_data", "industrial", "Tuna_fishing_density_noports.tif"))
```

Steps 3-6 are the steps to recreate each vessel type raster. Each raster needed to be reprojected to match the projections of the project area and rescaled to all be from 0-1 to have equally weight. However, `Prioritzr` does not like costs of zero, so we rescaled from 0.0001-1

#### **STEP 3: Recreate Longline Raster**
```{r}
### reproject the long-line raster to match the mz_rast_id
long_line_rp <- projectRaster(long_line, mz_rast_id, 
                              res = 10000, 
                              method = 'ngb',
                              crs = crs(mz_rast_id)) 

### Priortizr does not like cost values of NA so first let's change all the NA's --> 0
long_line_rp[is.na(long_line_rp[])] <- 0
plot(long_line_rp)

## let's check it!
long_line_df <- long_line_rp %>% 
  as.data.frame()

### since we changed all NAs to 0 we need to again mask it to change values on land and past the EEZ to NA
long_line_rp <- mask(long_line_rp, mz_eez_sf)
plot(long_line_rp)

long_line_df <- long_line_rp %>% 
  as.data.frame()

### rescale from 0-1
long_line_rescaled_df <- long_line_rp %>%
  as.data.frame() %>%
  mutate(longline_convert = case_when(fishing_line_kernel_density_noports < 0.4 ~ 0,
                                       TRUE ~ as.numeric(.$fishing_line_kernel_density_noports))) %>%
  mutate(longline_rescale = longline_convert/max(longline_convert, na.rm = TRUE)) %>%
  mutate(longline_nonzero = case_when(longline_rescale == 0 ~ .0001,
                                      TRUE ~ as.numeric(.$longline_rescale))) %>%
  mutate(id = 1:n())

### another way to rescale from 0 to 1
#long_line_rescaled <- log(long_line_rp+1)/ cellStats(log(long_line_rp+1), max)
# long_line_rescaled <- long_line_rp/maxValue(long_line_rp)
#
### let's check it again
# long_line_rescaled_df <- long_line_rescaled %>% 
#   as.data.frame()

long_line_rescaled <- subs(mz_rast_id, long_line_rescaled_df, by = 'id', which = 'longline_nonzero')

plot(long_line_rescaled)

```

#### **STEP 4: Recreate Gamba Raster**
```{r}
### reproject the gamba raster to match the mz_rast_id
gamba_rp <- projectRaster(gamba, mz_rast_id, 
                          res = 10000, 
                          method = 'ngb',
                          crs = crs(mz_rast_id))
plot(gamba_rp)

### Priortizr does not like cost values of NA so first let's change all the NA's --> 0
gamba_rp[is.na(gamba_rp[])] <- 0
plot(gamba_rp)

## let's check it!
gamba_df <- gamba_rp %>% 
  as.data.frame()

### since we changed all NAs to 0 we need to again mask it to change values on land and past the EEZ to NA
gamba_rp <- mask(gamba_rp, mz_eez_sf)
plot(gamba_rp)

### rescale from 0-1
gamba_rescaled_df <- gamba_rp %>%
  as.data.frame() %>%
  mutate(gamba_convert = case_when(Gamba_fishing_density_noports < .52 ~ 0,
                                       TRUE ~ as.numeric(.$Gamba_fishing_density_noports))) %>%
  mutate(gamba_rescale = gamba_convert/max(gamba_convert, na.rm = TRUE)) %>%
  mutate(gamba_nonzero = case_when(gamba_rescale == 0 ~ .0001,
                                      TRUE ~ as.numeric(.$gamba_rescale))) %>%
  mutate(id = 1:n())

### another way to rescale from 0 to 1
# #gamba_rescaled <- log(gamba_rp+1)/ cellStats(log(gamba_rp+1), max)
# gamba_rescaled <- gamba_rp/maxValue(gamba_rp)

### let's check it again
# gamba_rescaled_df <- gamba_rescaled %>% 
#   as.data.frame()

gamba_rescaled <- subs(mz_rast_id, gamba_rescaled_df, by = 'id', which = 'gamba_nonzero')

plot(gamba_rescaled)

```

#### **STEP 5: Recreate Shrimp Raster**
```{r}
### reproject the shrimp raster to match the mz_rast_id
shrimp_rp <- projectRaster(shrimp, mz_rast_id, 
                           res = 10000, 
                           method = 'ngb',
                           crs = crs(mz_rast_id))
plot(shrimp_rp)

### Priortizr does not like cost values of NA so first let's change all the NA's --> 0
shrimp_rp[is.na(shrimp_rp[])] <- 0
plot(shrimp_rp)

## let's check it!
shrimp_df <- shrimp_rp %>% 
  as.data.frame()

### since we changed all NAs to 0 we need to again mask it to change values on land and past the EEZ to NA
shrimp_rp  <- mask(shrimp_rp, mz_eez_sf)
plot(shrimp_rp)

### rescale from 0-1
shrimp_rescaled_df <- shrimp_rp %>%
  as.data.frame() %>%
  mutate(shrimp_convert = case_when(Shrimp_fishing_density_noports < 2.5 ~ 0,
                                       TRUE ~ as.numeric(.$Shrimp_fishing_density_noports))) %>%
  mutate(shrimp_rescale = shrimp_convert/max(shrimp_convert, na.rm = TRUE)) %>%
  mutate(shrimp_nonzero = case_when(shrimp_rescale == 0 ~ .0001,
                                      TRUE ~ as.numeric(.$shrimp_rescale))) %>%
  mutate(id = 1:n())

### another way to rescale from 0 to 1
# #shrimp_rescaled <- log(shrimp_rp+1)/ cellStats(log(shrimp_rp+1), max)
# shrimp_rescaled <- shrimp_rp/maxValue(shrimp_rp)

### let's check it again
# shrimp_rescaled_df <- shrimp_rescaled %>% 
#   as.data.frame()

# plot(shrimp_rescaled)

shrimp_rescaled <- subs(mz_rast_id, shrimp_rescaled_df, by = 'id', which = 'shrimp_nonzero')

plot(shrimp_rescaled)

```

#### **STEP 6: Recreate Tuna Raster**
```{r}
### reproject the tuna raster to match the mz_rast_id
tuna_rp <- projectRaster(tuna, mz_rast_id, 
                         res = 10000,
                         method = 'ngb',
                         crs = crs(mz_rast_id)) 
plot(tuna_rp)

### Priortizr does not like cost values of NA so first let's change all the NA's --> 0
tuna_rp[is.na(tuna_rp[])] <- 0
plot(tuna_rp)

## let's check it!
tuna_df <- tuna_rp %>% 
  as.data.frame()

### since we changed all NAs to 0 we need to again mask it to change values on land and past the EEZ to NA
tuna_rp  <- mask(tuna_rp, mz_eez_sf)
plot(tuna_rp)

### rescale from 0-1
tuna_rescaled_df <- tuna_rp %>%
  as.data.frame() %>%
  mutate(tuna_convert = case_when(Tuna_fishing_density_noports < 2.3 ~ 0,
                                       TRUE ~ as.numeric(.$Tuna_fishing_density_noports))) %>%
  mutate(tuna_rescale = tuna_convert/max(tuna_convert, na.rm = TRUE)) %>%
  mutate(tuna_nonzero = case_when(tuna_rescale == 0 ~ .0001,
                                      TRUE ~ as.numeric(.$tuna_rescale))) %>%
  mutate(id = 1:n())


### another way to rescale from 0 to 1
#long_line_rescaled <- log(long_line_rp+1)/ cellStats(log(long_line_rp+1), max)
# long_line_rescaled <- long_line_rp/maxValue(long_line_rp)
# 
# ## let's check it again
# long_line_rescaled_df <- long_line_rescaled %>% 
#   as.data.frame()

tuna_rescaled <- subs(mz_rast_id, tuna_rescaled_df, by = 'id', which = 'tuna_nonzero')

plot(tuna_rescaled)


```

#### **STEP 7: Create a raster stack**
Like the conservation features, the costs can only be one raster or a raster stack, so the next step is to create a raster stack off these four rasters. Then, rescale once again so it is 0-1 rather than 0-4. 

```{r}
## create raster stack of all industrial layers
industrial_stack <-  stack(long_line_rescaled, gamba_rescaled, shrimp_rescaled, tuna_rescaled)
plot(industrial_stack)

## let's check it
industrial_stack_df <- industrial_stack %>% 
  as.data.frame()

## use the raster::calc function to sum all layer into one
industrial_stack_sum <- calc(industrial_stack, fun = sum)
plot(industrial_stack_sum)

## rescale it from 0-1 so all cost layers are on the same scale
#industrial_stack_rescaled <- log(industrial_stack_sum+1)/ cellStats(log(industrial_stack_sum+1), max)
industrial_stack_sum_rescaled <- industrial_stack_sum/maxValue(industrial_stack_sum)
plot(industrial_stack_sum_rescaled)

## let's check it
industrial_stack_sum_rescaled_df <- industrial_stack_sum_rescaled %>% as.data.frame()

### save it
writeRaster(industrial_stack_sum_rescaled, here('set_up/final_rasters/costs/cost_industrial_sum.tif'), overwrite = TRUE)

### save the stack that's not rescaled a second time
writeRaster(industrial_stack_sum, here('set_up/final_rasters/costs/cost_industrial_sum_unscaled.tif'), options = "INTERLEAVE=BAND", overwrite = TRUE)

```


For fun... an interactive map of the final industrial fishing pressure raster!
```{r}
tmap_mode("view") # Set to interactive viewing

tm_shape(industrial_stack_sum_rescaled) +
  tm_basemap("Stamen.Terrain")+
  tm_raster("layer", alpha=0.9, palette = "viridis")
  
```
 
Outputs:

 - `cost_industrial_sum.tif` = raster of summed and rescaled industrial fishing pressure
 - `cost_industrial_sum_unscaled.tif` = raster of summed industrial fishing pressure (not rescaled)
 

Citations:
Vessel Monitoring System (VMS) data - Wildlife Conservation Society, 2020.


End script.
