==> set_up/7_artisanal_layer_setup.Rmd <==
---
title: "Artisanal Cost Layer"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(raster)
library(tidyverse)
library(sf)
library(here)
library(fasterize)
library(stringr)
library(janitor)
library(tmap)
```

# Cost: Artisanal Fishing Pressure
Another cost we considered was artisanal fishing pressure. To create proxy, we used district level fishing effort from 2017, measured in fishing days, and country-wide mean distance traveled by fishers to assign a fishing pressure value to each cell within the travel range. We set points at 1 km intervals along the entire coastline and for each district, divided the total effort by the number of coastal points within that district, so that each point along the coast had a fishing effort assigned to it. Next, we used the Kernel Density tool within ArcGIS with these inputs to create a search radius to distribute this fishing effort from each coastal point. Based on information provided by the Mozambique government, the average maximum distance traveled by artisanal fishers is 15-km. For each 1-km ocean cell in the output, the kernel density function searched within a 15-km radius and summed the values of any coastal district points. As such, the further an ocean cell was from shore, the fewer points were likely to be included, which is why the fishing effort values decreased further offshore. After 15-km to any shoreline, the fishing effort is zero. This Rmarkdown walks through how we wrangle this cost layer to be use in `Prioritzr`.

#### **STEP 1: Read in Mozambique EEZ and Shapefile**
```{r}
## read in MZ EEZ tif
mz_rast_id <- raster(here("set_up/final_rasters/mz_eez_templates/mz_rast_id.tif"))

## Grab Moz Exclusive Economic Zone (EEZ) shapefile from local computer
mz_eez_path <- 'G:/group_project/data/mz_eez'
mz_shape <- list.files(mz_eez_path, full.names = TRUE) 

mz_eez_sf <- sf::read_sf(mz_shape[6])

## check coordinate system
st_crs(mz_eez_sf)

## Add a buffer of 10 km to make sure cells on the outer edges of the EEZ are included
mz_eez_buffer_sf <- sf::st_buffer(mz_eez_sf, dist = 10000)
```

#### **STEP 2: Read in artisanal fisheries raster**
```{r}
### pull in the artisanal fishing pressure tif file from costs folder in repo
art_fish_rast <- raster(here("set_up", "raw_data", "artisanal", "kernel_d_fishing_moz_v2.tif"))
```


#### **STEP 3: Recreate Artisanal Raster**
This includse reprojecting the artisanal fishing raster to match crs of our planning units, removing zero values, and rescaling from 0-1 
```{r}
### now we need to grab the Moz raster we previously created to make sure the artisanal fishing raster is in the right projections
art_fish <- raster::projectRaster(art_fish_rast, mz_rast_id, 
                                  res = 10000, 
                                  method = 'ngb')

### create a data frame to see what value ranges we are working with
art_fish_df <- as.data.frame(art_fish)

### plot to see what it looks like
plot(art_fish)

### Priortizr does not like cost values of NA so first let's change all the NA's --> 0
art_fish[is.na(art_fish[])] <- 0
plot(art_fish)

### since we changed all NAs to 0 we need to again mask it to change values on land and past the EEZ to NA
art_fish <- mask(art_fish, mz_eez_sf)
plot(art_fish)

### now let's check it #0-1 method
art_fish_df <- art_fish %>% 
  as.data.frame() %>%
  mutate(artisanal_convert = case_when(kernel_d_fishing_moz_v2 < 0.1 ~ 0,
                                       TRUE ~ as.numeric(.$kernel_d_fishing_moz_v2))) %>%
  mutate(artisanal_rescale = artisanal_convert/max(artisanal_convert, na.rm = TRUE)) %>% 
  mutate(artisanal_nonzero = case_when(artisanal_rescale == 0 ~ .0001, 
                                      TRUE ~ as.numeric(.$artisanal_rescale))) %>% 
  mutate(id = 1:n())

### another way to rescale from 0 to 1
#art_fish_rescaled <- log(art_fish+1)/ cellStats(log(art_fish+1), max)
# art_fish_rescaled <- art_fish/maxValue(art_fish)
# plot(art_fish_rescaled)

### let's check it again!
# art_fish_rescaled_df <- art_fish_rescaled %>% 
#   as.data.frame() 

### now that we've changed the zeros and rescaled, convert back to a raster
art_fish_rescaled <- subs(mz_rast_id, art_fish_df, by = 'id', which = 'artisanal_nonzero')
plot(art_fish_rescaled)

## Looks good lets save as a tif
writeRaster(art_fish_rescaled, here('set_up/final_rasters/costs/cost_art_fish.tif'), overwrite = TRUE)
```


For fun... an interactive map of the final artisanal fishing pressure raster!
```{r}
tmap_mode("view") # Set to interactive viewing

tm_shape(art_fish_rescaled) +
  tm_basemap("Stamen.Terrain")+
  tm_raster("artisanal_nonzero", alpha=0.9, palette = "viridis")
  
```

Outputs:

 - `cost_art_fish.tif` = raster of rescaled artisanal fishing pressure
 
Citations:
Boletim Estatístico da Pesca e Aquacultura (2006 - 2017) - Ministério do Mar, Águas Interiores e Pescas, Rua Marquês de Pombal, Nº 258, Caixa Postal 1723, Maputo, Telefone: + 258-21357100, Celular: 82-3073530, Website: www.mozpescas.gov.mz, Correio electrónico: mpescas@mozpescas.gov.mz

Fishing Effort at 10th Degree - Global Fishing Watch. 2020. Accessed on 09/20/2020. https://globalfishingwatch.org/data-download/datasets/public-fisshing-effort-10:v20200316.


End script.
