==> set_up/8_combined_fishing_layer_setup.Rmd <==
---
title: "Combined Fishing Pressure"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(raster)
library(tidyverse)
library(sf)
library(here)
library(fasterize)
library(stringr)
library(janitor)
library(tmap)
```

# Cost: Combined Fishing Pressure
To minimize the impacts and conflicts that marine protected areas have on all fishers in Mozambique’s EEZ, artisanal and industrial, we created a combined fishing pressure layer. The Mozambique government has expressed their intention to minimize the impact of MPAs on fishers, presumably both artisanal and industrial, and therefore avoid conflict resulting from MPAs placed in areas with high fishing pressure. Since the aim of this layer is to prioritize outputs that avoid overlap with high fishing pressure areas, artisanal and industrial fishers were weighted equally. We combined the artisanal and industrial fishing pressure layers and rescaled the cumulative layer between 0 and 1. 


#### **STEP 1: Read in Mozambique EEZ and Shapefile**
```{r}
## read in MZ EEZ tif
mz_rast_id <- raster(here("set_up/final_rasters/mz_eez_templates/mz_rast_id.tif"))

## Grab Moz Exclusive Economic Zone (EEZ) shapefile from local computer
mz_eez_path <- 'G:/group_project/data/mz_eez'
mz_shape <- list.files(mz_eez_path, full.names = TRUE) 

mz_eez_sf <- sf::read_sf(mz_shape[6])

## check coordinate system
st_crs(mz_eez_sf)

## Add a buffer of 10 km to make sure cells on the outer edges of the EEZ are included
mz_eez_buffer_sf <- sf::st_buffer(mz_eez_sf, dist = 10000)
```


#### **STEP 2: Read in Artisanal and Industrial Fishing Rasters**
```{r}
### read in artisanal fishing raster
art_fish_rast <- raster(here("set_up", "final_rasters", "costs", "cost_art_fish.tif"))

### read in industrial (rescaled and summed) raster
industrial_sum_rescaled <- raster(here("set_up", "final_rasters", "costs", "cost_industrial_sum.tif"))
```


#### **STEP 4: Create raster stack of all fishing**
`Prioritzr` only likes one cost layer, so we create a raster stack of all fishing.
```{r}
### create raster stack
combined_fishing_stack <- stack(industrial_sum_rescaled, art_fish_rast)

### plot it!
plot(combined_fishing_stack)

combined_stack_sum <- calc(combined_fishing_stack, fun = sum)
plot(combined_stack_sum)

### save sum 
writeRaster(combined_stack_sum, here('set_up/final_rasters/costs/combined_stack_sum.tif'), 
            overwrite = TRUE)

### save stack
writeRaster(combined_fishing_stack, here('set_up/final_rasters/costs/combined_fishing_stack.tif'), 
            options="INTERLEAVE=BAND",
            overwrite = TRUE)
```

For even more fun... an interactive map of combined fishing pressure!
```{r}
tmap_mode("view") # Set to interactive viewing

tm_shape(combined_stack_sum) +
  tm_basemap("Stamen.Terrain")+
  tm_raster("layer", alpha=0.9, palette = "viridis")
  
```

# Cost: Bell Curve
We started exploring looking at the bell curve approach to find a middle ground between fishing pressure and conservation features. Based on our combined fishing pressure cost layer, we found the middle values (0.1-0.2) and made them the least costly so they would be prioritized in our model. We did not run this analysis using `Prioritzr` and it was not presented in our report due to time constraint. However, it was fun thinking about an alternative approach that could maximize protection for shark and rays as well as keep the local fishermen's most productive fishing grounds. 


#### Exploring the bell curve
```{r}
### create plot that matches arcgis

### convert raster to df
combined_stack_sum_df <- rasterToPoints(combined_stack_sum) %>% 
  as.data.frame()

### plot it like the final version in arcgis
ggplot()+
  geom_raster(data = combined_stack_sum_df, aes(x,y, fill = layer))+
  geom_sf(data = mz_eez_sf, fill = NA)+
  scale_fill_continuous(type = "viridis",breaks =  c(0,0.02,0.1,0.2,.44,1))+
  theme_minimal()

```
#### Reclassifying the values
Made the middle ground values (0.1-0.2) or yellow in our Arc Maps the lowest values (least costly)
```{r}
m <- c(0.00, 0.02, 4,    
       0.02, 0.1, 2,    
       0.1, 0.2, 1,
       0.2, 0.44, 3,
       0.44, 1.00, 5)

### wrap the vector into a matrix by rows
rcl_mtx <- matrix(m, ncol=3, byrow=TRUE)

### use raster::reclassify() to reclassify the values 
combined_reclass <- reclassify(combined_stack_sum, rcl_mtx)
plot(combined_reclass)

### check if it worked
combined_reclass_df <- rasterToPoints(combined_reclass) %>% 
  as.data.frame()

### plot it! the darkest color should be the least costly
ggplot()+
  geom_raster(data = combined_reclass_df, aes(x,y, fill = layer))+
  geom_sf(data = mz_eez_sf, fill = NA)+
  scale_fill_continuous(type = "viridis")

writeRaster(combined_reclass, here('set_up/final_rasters/costs/combined_reclass.tif'), 
            overwrite = TRUE)
```
#### Distance Decay
Using the least costly areas found above, we used the `raster::distance` to create a distance decay. This means that the further away from the least costly area, the greater the cost.
```{r}
### creating a matrix that makes all values beside 0.1-0.2 (middle ground area) NA -- all other values need to be NA for distance decay to work
m_dist <- c(0.00, 0.02, NA,    
       0.02, 0.1, NA,    
       0.1, 0.2, 1,
       0.2, 0.44, NA,
       0.44, 1.00, NA)

### wrap the vector into a matrix by rows
rcl_mtx_dist <- matrix(m_dist, ncol=3, byrow=TRUE)

### reclassify values
combined_reclass_dist <- reclassify(combined_stack_sum, rcl_mtx_dist)
plot(combined_reclass_dist)

## use the raster::distance() function, which calculates a distance for every NA cell to the nearest non-NA cell.
combined_reclass_dist_rast <- raster::distance(combined_reclass_dist) %>% 
  mask(mz_eez_sf) # mask so the land distances are not calculated

## plot it!
plot(combined_reclass_dist_rast)

### save it
writeRaster(combined_dist_rast, here('set_up/final_rasters/costs/combined_reclass_dist_rast.tif'), overwrite = TRUE)
```

Outputs:

 - `combined_stack_sum.tif` = raster of summed combined fishing pressure
 - `combined_fishing_stack.tif`= raster of stacked combined fishing pressure
 - `combined_reclass.tif` = raster of reclassifed combined fishing pressure (bell-curve)
 - `combined_reclass_dist_rast.tif`= distance decay raster of reclassified combined fishing pressure (bell-curve)

End script.
