==> set_up/9_area_cost_layer_setup.Rmd <==
---
title: "Area Cost Layer Setup"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(raster)
library(tidyverse)
library(sf)
library(here)
library(fasterize)
library(stringr)
library(janitor)
library(tmap)
```

# Cost: Area
To explore an optimal conservation design that does not consider costs to specific stakeholders we used planning unit area as a surrogate for cost. Since our planning units are uniform in size, we set the cost to be uniform across our entire project area. This uniform cost will not give us useful solutions that minimize a cost to any stakeholder, however, we can use this as a baseline scenario against which our other outputs could be compared as we add more complex costs.

#### **STEP 1: Read in Mozambique EEZ and Shapefile**
```{r}
## read in MZ EEZ tif
mz_rast_id <- raster(here("set_up/final_rasters/mz_eez_templates/mz_rast_id.tif"))

## Grab Moz Exclusive Economic Zone (EEZ) shapefile from local computer
mz_eez_path <- 'G:/group_project/data/mz_eez'
mz_shape <- list.files(mz_eez_path, full.names = TRUE) 

mz_eez_sf <- sf::read_sf(mz_shape[6])

## check coordinate system
st_crs(mz_eez_sf)

## Add a buffer of 10 km to make sure cells on the outer edges of the EEZ are included
mz_eez_buffer_sf <- sf::st_buffer(mz_eez_sf, dist = 10000)
```

#### **STEP 2: Assign raster values**
```{r}
### pull in the initial MZ EEZ raster to start with
area_raster <- raster(here("set_up", "final_rasters", "mz_eez_templates", "mz_rast_id.tif"))

### assign a uniform cost value of 0.5 to all cells since the area of the planning units are the same. 
### We assigned a value of 0.5 since we will transform cost layers in future simulations to be between 0-1 and this 0.5 is a middle value in this range 
values(area_raster) <- 0.5

### mask the area raster to just the MZ EEZ, so all cells outside the EEZ are assigned NA AND rename it to cost_area
cost_area <- mask(area_raster, mz_rast_id)

### check it by plotting!
plot(cost_area)

### write raster
writeRaster(cost_area, here('set_up/final_rasters/costs/area_cost.tif'), overwrite = TRUE)
```

Another interactive map!
```{r}
tmap_mode("view") # Set to interactive viewing

tm_shape(cost_area) +
  tm_basemap("Stamen.Terrain")+
  tm_raster("mz_rast_id",alpha=0.9, palette="viridis")
  
```

Outputs:

 - `area_cost.tif` = raster of area cost layer

End script.
